Rename: 디렉토리 구조 수정
팀 디렉토리 컨벤션 수정에 따른 디렉토리 구조 수정
