--- matjalalzz/src/main/java/shop/matjalalzz/domain/user/dto/OAuthSignUpRequest.java
@@ -1,23 +1,0 @@
-package shop.matjalalzz.domain.user.dto;
-
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import shop.matjalalzz.domain.user.entity.enums.Gender;
-
-public record OAuthSignUpRequest(
-	@NotBlank(message = "닉네임은 필수 입력값입니다.")
-	String nickname,
-
-	@NotBlank(message = "전화번호는 필수 입력값입니다.")
-	String phoneNumber,
-
-	@NotBlank(message = "이름은 필수 입력값입니다.")
-	String name,
-
-	@Min(value = 0, message = "나이는 0 이상이어야 합니다.")
-	long age,
-
-	@NotNull(message = "성별은 필수 입력값입니다.")
-	Gender gender
-) {}
