--- matjalalzz/src/main/java/shop/matjalalzz/global/unit/AuditorAwareImpl.java
@@ -1,23 +1,0 @@
-package shop.matjalalzz.global.unit;
-
-import java.util.Optional;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AuditorAwareImpl implements AuditorAware<String> {
-
-    @Override
-    public Optional<String> getCurrentAuditor() {
-        // 예시: Spring Security에서 현재 로그인한 사용자 이름 반환
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null || !authentication.isAuthenticated()) {
-            return Optional.empty();
-        }
-
-        return Optional.of(authentication.getName()); // 또는 userId
-    }
-}
