+++ matjalalzz/src/main/java/shop/matjalalzz/user/entity/enums/Gender.java
@@ -1,0 +1,5 @@
+package shop.matjalalzz.user.entity.enums;
+
+public enum Gender {
+    M, W
+}
