+++ matjalalzz/src/main/java/shop/matjalalzz/user/entity/enums/Role.java
@@ -1,0 +1,6 @@
+package shop.matjalalzz.user.entity.enums;
+
+public enum Role {
+    ADMIN, USER, OWNER
+
+}
